Atualização de script.

+++ excel/deletes/1 - delete method/script.py
@@ -3,7 +3,9 @@
 # Altere o nome das variáveis para funcionar corretamente em seu script.
 
 # Importação de biblioteca.
-import os  # Biblioteca para manipular arquivos do sistema operacional do windows.
+import os  # Módulo para fornecer acesso a funções específicas do sistema para lidar com o sistema de arquivos, processos, planejador, etc.
+import sys  # Módulo para fornecer funções e variáveis usadas para manipular diferentes partes do ambiente de tempo de execução do Python e apesar de serem completamente diferentes, muitas pessoas confundem o módulo sys e o módulo os (módulo para manipular o sistema operacional).
+import time  # Módulo provê várias funções relacionadas à tempo, onde neste caso estamos utilizando para a função (sleep).
 
 # Declaração do caminho origem, ou seja, onde os arquivos se encontram.
 caminho_origem = r'C:\Windows\Temp\testes'
@@ -11,27 +13,47 @@ caminho_origem = r'C:\Windows\Temp\testes'
 # Declaração de variável com o nome do arquivo, que neste caso irá facilitar para utilização no decorrer do script.
 nome_arquivo = 'Sou-Milionário.xlsx'
 
-# Loop para verificar os arquivos existentes no caminho origem, ou seja, na variável que carrega o caminho.
-for root, dirs, files in os.walk(caminho_origem):
+# Variável com o comando para verificar se o repositório origem, o mesmo da variável (source_repository_path) existe.
+verificao_existencia_repositorio = os.path.isdir(caminho_origem)
 
-    # Este (if) verifica se o (local - root) é o mesmo do repositório (nome_caminho), visto que os arquivos utilizados estão neste repositório.
-    # Também certifica que o script não percorra outros repositórios dentro do repositório principal da variável (caminho_origem).
-    if root == caminho_origem:
+# Caso a cláusula seja verdadeira, então irá executar o comando abaixo.
+# Verifica se o repositório origem existe. Caso existe, irá para o comando abaixo.
+if verificao_existencia_repositorio:
 
-        # Loop para verificar somente os arquivos do for acima.
-        for file in files:
+    # Loop para verificar os arquivos existentes no caminho origem, ou seja, na variável que carrega o caminho.
+    for root, dirs, files in os.walk(caminho_origem):
 
-            # If para verificar somente o arquivo com o noem da variável (nome_arquivo).
-            if file == nome_arquivo:
+            # Este (if) verifica se o (local - root) é o mesmo do repositório (nome_caminho), visto que os arquivos utilizados estão neste repositório.
+            # Também certifica que o script não percorra outros repositórios dentro do repositório principal da variável (caminho_origem).
+            if root == caminho_origem:
 
-                # Caso o arquivo da variável (nome_arquivo) seja localizado, então o processo irá prosseguir e ignorar este arquivo.
-                continue
-            
-            # Else para excluir os demais arquivos que estiverem no caminho origem, ou seja, na variável que carrega o caminho.
-            else:
+                # Loop para verificar somente os arquivos do for acima.
+                for file in files:
 
-                # Remove os arquivos que não tem o nome da variável (nome_arquivo).
-                os.remove(caminho_origem + '\\' + file)
+                    # If para verificar somente o arquivo com o noem da variável (nome_arquivo).
+                    if file == nome_arquivo:
 
-                # Imprimi na tela a mensagem dos arquivos que foram excluídos.
-                print('Arquivo', file, 'removido com sucesso!')
+                        # Caso o arquivo da variável (nome_arquivo) seja localizado, então o processo irá prosseguir e ignorar este arquivo.
+                        continue
+                    
+                    # Else para excluir os demais arquivos que estiverem no caminho origem, ou seja, na variável que carrega o caminho.
+                    else:
+
+                        # Remove os arquivos que não tem o nome da variável (nome_arquivo).
+                        os.remove(caminho_origem + '\\' + file)
+
+                        # Imprimi na tela a mensagem dos arquivos que foram excluídos.
+                        print('Arquivo', file, 'removido com sucesso!')
+
+# Este (else) será executado caso a cláusula do (if) seja falsa.
+# Se a cláusula for falsa, então quer dizer que o repositório não existe, logo haverá mensagem abaixo.
+else:
+
+    # Mensagem ao usuário caso o repositório não exista.
+    print('Repositório (' + caminho_origem + ') não existe. \nPor favor, verifique o caminho informado ou se o repositório existe e tente novamente. \nEssa tela será fechada em 10 segundos. \nProcesso finalizado sem êxito.')
+
+    # Pausar ou colocar para dormir a execução do script por 30 segundos até a execução do comando abaixo.
+    time.sleep(10)
+
+    # Caso ocorra o erro, então o script é encerrado e não prosseguirá para o próximo passo, assim evitando erros para o usuário posteriormente.
+    sys.exit()
